test(e2e): split endpoint creation into it blocks; add 30s wait to enableExternalProviders
testGenAiSettings.cy.ts:
- Move custom endpoint UI creation to its own it block
- Move Add to playground + OGX/LSD infrastructure wait to a second it block
- retryableBefore now only handles cluster setup and session
- Remove promptManagement flag from RAG it block
- Remove guardrails flag from guardrails it block

testGenAiModels.cy.ts:
- Remove custom endpoint it block — coverage now in testGenAiSettings

enableExternalProviders():
- Add 30 s wait after cluster poll so the backend ResourceWatcher
  propagates before UI validation fires
